Fix compilation error

+++ 0x0B-more_malloc_free/1-string_nconcat.c
@@ -34,7 +34,7 @@ char *string_nconcat(char *s1, char *s2, unsigned int n)
 	for (i = 0; s1[i] != '\0'; i++)
 		ptr[i] = s1[i];
 	j = i;
-	for (i = 0; j < len; j++)
+	for (i = 0; j < (int)len; j++)
 	{
 		ptr[j] = s2[i];
 		i++;
